move download button
